+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.3.0-rc4
+
+Added _enyo.perfNow()_, a high-precision, high performance monotonic timestamp, which is independent of changes
+to the system clock and safer for use in animation, etc.  Falls back to _enyo.now()_ (based on the JS _Date_
+object) on platforms where `window.performance.now()` is not available.
+
+_enyo.Control.getAbsoluteShowing()_ now accepts an optional boolean parameter that, if `true`, will
+skip the _getBounds()_ call internally and rely solely on the return value of the _getShowing()_
+public API.
+
+_enyo.Scroller_ has been updated such that it will only propagate _resize_ events to children when its'
+_showing_ property is `true`.
+
 ## 2.3.0-rc1
 
 _enyo.dispatcher.capture_ API no longer bubbles all captured events through the normal event chain, but